requirements.txt
arrow==0.14.7
asgiref==3.2.7
astroid==2.3.3
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-cogwheels==0.3
django-modelcluster==5.0.1
django-recaptcha==2.0.6
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
ics==0.7
idna==2.9
isort==4.3.21
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.2
psycopg2==2.8.4
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
TatSu==4.4.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
uWSGI==2.0.18
wagtail==2.8
wagtail-django-recaptcha==1.0
wagtailmenus==3.0.1
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
wrapt==1.11.2
